[change] Author name change
Litrain and m-h-san are the same person. It can continue to be used under the MIT License as before.

This can be confirmed from the fact that https://github.com/m-h-san/m3ripple redirects to https://github.com/yuyake-litrain/m3ripple.

If you have any other questions or concerns about the license, feel free to ask.